[APPC-192] - appcoins-sdk is now include appcoins-sdk-iab.
