+++ appcoins-sdk-iab/proguard-rules.pro
