+++ appcoins-sdk-iab/src/main/AndroidManifest.xml
