+++ appcoins-sdk-iab/src/main/res/values/strings.xml
