--- appcoins-sdk/src/main/java/com/asf/appcoins/sdk/payment/PaymentStatus.java
@@ -1,24 +1,0 @@
-package com.asf.appcoins.sdk.payment;
-
-import com.asf.appcoins.sdk.entity.Transaction.Status;
-
-/**
- * Created by neuro on 13-03-2018.
- */
-
-public enum PaymentStatus {
-  SUCCESS, FAIL, PENDING;
-
-  public static PaymentStatus from(Status status) {
-    switch (status) {
-      case FAILED:
-        return FAIL;
-      case PENDING:
-        return PENDING;
-      case ACCEPTED:
-        return SUCCESS;
-      default:
-        throw new IllegalArgumentException("Status unknown");
-    }
-  }
-}
