--- appcoins-sdk/src/test/java/com/asf/appcoins/sdk/ExampleUnitTest.java
@@ -1,16 +1,0 @@
-package com.asf.appcoins.sdk;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-  @Test public void addition_isCorrect() throws Exception {
-    assertEquals(4, 2 + 2);
-  }
-}
