+++ toolbox/build.gradle
@@ -37,5 +37,5 @@ dependencies {
 
   testImplementation lib.junit
 
-  api project(':appcoins-sdk')
+  api project(':appcoins-sdk-iab')
 }
